change vartype to 64 bits

+++ src/ducky.h
@@ -12,7 +12,7 @@ void ducky_main(int fd, bool verbose);
 void ducky_compile(int fd, bool verbose, int out_fd);
 void ducky_vm(int fd);
 
-typedef int32_t imm_t;
+typedef int64_t imm_t;
 typedef uint8_t instr_t;
 typedef uint16_t varid_t;
 typedef imm_t vartype;
